Remove core package
