--- src/main/java/com/hotels/hcommon/hive/metastore/core/compatibility/HiveMetaStoreClientCompatibility.java
@@ -1,27 +1,0 @@
-/**
- * Copyright (C) 2017-2018 Expedia Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.hotels.hcommon.hive.metastore.core.compatibility;
-
-import org.apache.hadoop.hive.metastore.api.MetaException;
-import org.apache.hadoop.hive.metastore.api.NoSuchObjectException;
-import org.apache.hadoop.hive.metastore.api.Table;
-import org.apache.thrift.TException;
-
-public interface HiveMetaStoreClientCompatibility {
-
-  Table getTable(String dbname, String name) throws MetaException, TException, NoSuchObjectException;
-
-}
